Extract logic for changing user role into a separate ChangeUserRoleDialog component
